modularize semantic analyzer
