+++ src/parser/typing.py
@@ -3,13 +3,13 @@ from lexer.token import Token
 from lexer.tokens import TokenType
 from syntax.ast import *
 
-class ParserABC():
+class ParserABC(Protocol):
     def parse(self) -> Program: ...
     def consume(self, token_type: TokenType) -> Token: ...
     def current(self) -> Token: ...
     def is_eof(self) -> bool: ...
 
-class StatementParserABC():
+class StatementParserABC(Protocol):
     def parse_statements(self) -> List[Statement]: ...
     def parse_statement(self) -> Statement: ...
     def parse_expression_statement(self) -> ExpressionStatement: ...
@@ -23,10 +23,10 @@ class StatementParserABC():
     def parse_function_declaration(self) -> FunctionDeclaration: ...
     def parse_echo_statement(self) -> EchoStatement: ...
 
-class ExpressionParserABC():
+class ExpressionParserABC(Protocol):
     def parse_expression(self) -> Expression: ...
     def parse_assignment_expression(self) -> Expression: ...
-    def parse_binary_expression(self, precedence: int = 0) -> Expression: ...
+    def parse_binary_expression(self, precedence: int) -> Expression: ...
     def parse_unary_expression(self) -> Expression: ...
     def parse_primary_expression(self) -> Expression: ...
     def parse_identifier_expression(self) -> Expression: ...
