+++ src/semantic/symbol.py
@@ -1,8 +1,9 @@
 from typing import *
 from semantic.types import FunctionType, VarType
+from semantic.typing import ScopeABC, SymbolABC, SymbolTableABC
 from syntax.ast import EachStatement, FunctionDeclaration, RangeStatement, Statement, WhileStatement
 
-class Symbol:
+class Symbol(SymbolABC):
     """
     Represents a symbol in the symbol table.
 
@@ -17,7 +18,7 @@ class Symbol:
     def __repr__(self) -> str:
         return f'Symbol(name={self.name}, var_type={self.var_type})'
 
-class Scope:
+class Scope(ScopeABC):
     """
     Represents a scope in the symbol table.
 
@@ -26,14 +27,14 @@ class Scope:
         function_type (Optional[FunctionType]): The function type if this scope is a function scope, otherwise None.
     """
     def __init__(self, parent_node: Optional[Statement] = None) -> None:
-        self.symbols: Dict[str, Symbol] = {}
+        self.symbols: Dict[str, SymbolABC] = {}
         self.parent_node = parent_node
         self.reachable = True
 
     def __repr__(self) -> str:
         return f'Scope(parent_node={self.parent_node}, symbols={list(self.symbols.keys())})'
 
-class SymbolTable:
+class SymbolTable(SymbolTableABC):
     """
     Represents a symbol table for managing variables and functions.
 
@@ -42,7 +43,7 @@ class SymbolTable:
     """
     def __init__(self) -> None:
         """Initializes the symbol table with an empty global scope."""
-        self.scopes: List[Scope] = [Scope()]
+        self.scopes: List[ScopeABC] = [Scope()]
 
     def enter_scope(self, parent_node: Optional[Statement] = None) -> None:
         """Enters a new scope, optionally as a function scope.
@@ -78,7 +79,7 @@ class SymbolTable:
             raise KeyError(f'Symbol {name} already declared in the current scope')
         current_scope.symbols[name] = Symbol(name, var_type)
 
-    def lookup(self, name: str, limit_to_function: bool = False) -> Optional[Symbol]:
+    def lookup(self, name: str, limit_to_function: bool = False) -> Optional[SymbolABC]:
         """Looks up a symbol by name, starting from the innermost scope.
 
         Args:
@@ -94,7 +95,7 @@ class SymbolTable:
             if limit_to_function and isinstance(scope.parent_node, FunctionDeclaration):
                 break
 
-    def get_scope(self, symbol: Symbol) -> Optional[Scope]:
+    def get_scope(self, symbol: SymbolABC) -> Optional[ScopeABC]:
         """Gets the scope containing the given symbol.
 
         Args:
